Make clean command Mac friendly

+++ scripts/clean.js
@@ -4,7 +4,7 @@ const execCommand = require('./util/execCommand');
 const os = require('os');
 const platform = os.platform();
 
-if (platform.indexOf('win') !== -1) {
+if (platform.indexOf('win') !== -1 && platform.indexOf('darwin') === -1) {
   execCommand('rmdir /s /q node_modules');
   // not sure what other windows clean commands are
 } else {
